autogen.sh: Don't remove Makefile when NOREDIRECTMAKEFILE

+++ autogen.sh
@@ -63,7 +63,9 @@ if test -n "${NOCONFIGURE:-}"; then
 fi
 
 cd $olddir
-rm -f $srcdir/Makefile
+if [ -z "${NOREDIRECTMAKEFILE:-}" ]; then
+    rm -f $srcdir/Makefile
+fi
 $srcdir/configure --enable-maintainer-mode ${AUTOGEN_CONFIGURE_ARGS:-} "$@" || exit $?
 
 # Put a redirect makefile here
